AppModel refactor: Lets break up AppModel.js with 1444 lines into 8 focused modules.
